refactor(shared): fold payment_anomaly_* enums into enums.ts SSoT (#1383)
The payment_anomaly_kind / payment_anomaly_resolution value sets were
declared twice: inline literals fed to pgEnum in db/payment.ts, and again
as PAYMENT_ANOMALY_KINDS / PAYMENT_ANOMALY_RESOLUTIONS in
types/payment-anomaly.ts, linked only by a mirror test. This was the last
domain enum not sourced from the enums.ts SSoT.

- Move both arrays (+ derived types) into enums.ts, grouped with the other
  payment_* enums.
- Feed them to pgEnum() in db/payment.ts (identical values/order → no SQL).
- Re-export from types/payment-anomaly.ts so the web boundary import path
  (@kuruma/shared/types/payment-anomaly) stays stable — web reads them as
  pure data since it cannot import the Drizzle schema.
- Pin both arrays in enums.test.ts (the mirror test is now tautological, so
  the literal pin is the real fence for the enums.ts values).

db:generate produces zero diff; db:verify green. No migration.

Part of #1369.

Closes #1383
