+++ packages/shared/src/types/payment-anomaly.ts
@@ -14,25 +14,24 @@
  * Stripe sent a malformed event.
  */
 
-/** The `payment_anomaly_kind` DB enum, mirrored for the web boundary. Order must
- *  match the schema enum — pinned by `tests/types/payment-anomaly.test.ts`. */
-export const PAYMENT_ANOMALY_KINDS = ['DOUBLE_PAYMENT', 'AMOUNT_MISMATCH'] as const
+// The `payment_anomaly_kind` / `payment_anomaly_resolution` enums live in the
+// `enums.ts` SSoT (fed to `pgEnum` in `db/payment.ts`, pinned by `enums.test.ts`).
+// Re-exported here so the web boundary import path
+// (`@kuruma/shared/types/payment-anomaly`) stays stable — web cannot import the
+// Drizzle schema, so it reads these unions as pure data from this subpath.
+import {
+  PAYMENT_ANOMALY_KINDS,
+  PAYMENT_ANOMALY_RESOLUTIONS,
+  type PaymentAnomalyKind,
+  type PaymentAnomalyResolution,
+} from '../enums'
 
-export type PaymentAnomalyKind = (typeof PAYMENT_ANOMALY_KINDS)[number]
-
-/** The `payment_anomaly_resolution` DB enum, mirrored for the web boundary (#1075
- *  slice 3). An admin clears the review queue by tagging WHY a flagged charge is
- *  closed — it carries no money: `BENIGN` (not actually a problem), `INVESTIGATED`
- *  (looked into, no action), `REFUNDED_EXTERNALLY` (refunded in the Stripe
- *  dashboard, since in-app refund is deferred to a future slice). Order must match
- *  the schema enum — pinned by `tests/types/payment-anomaly.test.ts`. */
-export const PAYMENT_ANOMALY_RESOLUTIONS = [
-  'BENIGN',
-  'INVESTIGATED',
-  'REFUNDED_EXTERNALLY',
-] as const
-
-export type PaymentAnomalyResolution = (typeof PAYMENT_ANOMALY_RESOLUTIONS)[number]
+export {
+  PAYMENT_ANOMALY_KINDS,
+  PAYMENT_ANOMALY_RESOLUTIONS,
+  type PaymentAnomalyKind,
+  type PaymentAnomalyResolution,
+}
 
 /** One unresolved anomaly. Identifiers are carried so an admin can reconcile or
  *  refund: `stripeEventId` is the reconciliation handle, `stripePaymentIntentId`
